feat: unit test

+++ src/ast/syntax/param_list.cc
@@ -38,7 +38,7 @@ namespace kisyshot::ast::syntax {
 
     std::string ParamDeclaration::toString() const {
         std::string s = type->toString() + " " + name->toString();
-        for (int i = 0; i < dimension; ++i) {
+        for (size_t i = 0; i < dimension; ++i) {
             s += "[]";
         }
         return s;
